forked from main project

+++ java/java-tests/testData/inspection/inefficientStreamCount/afterStreamCount.java
@@ -1,10 +1,10 @@
-// "Replace Collection.stream().count() with Collection.size()" "true"
+// "Replace 'Collection.stream().count()' with 'Collection.size()'" "true"
 
 import java.util.Arrays;
 
 class Test {
   long cnt() {
       /*count*/
-      return (long) Arrays.asList('d', 'e', 'f')./*stream*/size()/*after*/;
+      return Arrays.asList('d', 'e', 'f')./*stream*/size()/*after*/;
   }
 }
