+++ java/java-tests/testData/inspection/inefficientStreamCount/afterStreamCountField.java
@@ -1,8 +1,8 @@
-// "Replace Collection.stream().count() with Collection.size()" "true"
+// "Replace 'Collection.stream().count()' with 'Collection.size()'" "true"
 
 import java.util.Arrays;
 
 class Test {
     /*count*/
-    long cnt = (long) Arrays.asList('d', 'e', 'f')./*stream*/size()/*after*/;
+    long cnt = Arrays.asList('d', 'e', 'f')./*stream*/size()/*after*/;
 }
